Project Euler Problem 104, O(n) solution - 2 sec
Linear solution to finding if the front and back of a fibonnaci number were pandigital. 

Commented out in the Calc Fib part, there is a O(LogN) solution to calculating a fiboncci number. The problem is it creates an overal problem complexity of O(NlogN) however in reality the worst case scenario won't happen because not every number is Pandigital for the last numbers. So I personally believe that this method will be faster in general, even though it is in theory a worse complexity.